Update to latest Foundry Portal & SDK
Update the sample to use the latest SK, Agents SDK & Foundry SDK to use the Project Endpoint instead of the previous connection string.

Adding much more details on how to use the sample.

Adding a full tutorial on how to create the Stocks Agent in Azure AI Foundry Agent Service.